refactor(streaming): remove AggExecutorWrapper
